Use SimpleMailer.test_mode! to set the test mode, bump version to 1.1.0
This changes the implementation so that the simple mailer test mode
is a global setting.  This works much better if SimpleMailer is
included in a class, as otherwise the test mode has to be set on
every instance of the class.

This commit also changes the test mode so that a single global
array is used to store test messages, instead of using an array
inside each instance of a class that includes SimpleMailer.